Add UnitRing Fields
