+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "engage"
-version = "0.6.5"
+version = "0.7.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
